Ensure that SJConf::conf_typed() only returns built-in python types, closes: #5603
conf_typed() already converts Conf and ConfSection objects to plain-old
dictionaries. The last step is to convert SJConf's List and Sequence types
to regular lists since the ConversionList object is a PITA to serialize
to pickle or json.

On paper, this breaks the SJConf public API, but I consider it a bug
that this function let ConversionList objects out in the wild.